created .m file, transplanted from master's thesis
The included MATLAB code list two functions (`TaylorSynth()` and `HalfPowerPt()`) that create a Taylor Illumination by leveraging the matrix math capabilities of MATLAB to simplify code and streamline calculation.  The functions, when used together (`TaylorSynth()` is the parent function to `HalfPowerPt()`) and accepting the inputs of theta3db, slldb, f0 and theta0, are able to create a Taylor Illumination.